removed bincode for rkyv

+++ crates/libmarathon/src/platform/desktop/executor.rs
@@ -179,8 +179,8 @@ impl AppHandler {
 
         // Create window entity with all required components (use logical size)
         // Convert physical pixels to logical pixels using proper floating-point division
-        let logical_width = (physical_size.width as f64 / scale_factor) as f32;
-        let logical_height = (physical_size.height as f64 / scale_factor) as f32;
+        let logical_width = (physical_size.width as f64 / scale_factor) as u32;
+        let logical_height = (physical_size.height as f64 / scale_factor) as u32;
 
         let mut window = bevy::window::Window {
             title: "Marathon".to_string(),
